Move base_envs.py, remove print
