+++ src/seals/util.py
@@ -110,7 +110,6 @@ def sample_distribution(
     p: np.ndarray, random: Optional[np.random.RandomState] = None,
 ) -> int:
     """Samples an integer with probabilities given by p."""
-    print(type(random))
     if random is None:
         random = np.random
     return random.choice(np.arange(len(p)), p=p)
